feat: add logic

+++ src/Lab5/Lab5.Application.Models/Operation/Operation.cs
@@ -1,3 +1,3 @@
 ﻿namespace Lab5.Application.Models.Operation;
 
-public record Operation(float AmountBefore, float Change, long OperationId);
+public record Operation(double AmountBefore, double Change, long OperationId);
