+++ src/Lab5/Lab5.Application.Models/User/User.cs
@@ -1,3 +1,3 @@
 ﻿namespace Lab5.Application.Models.User;
 
-public record User(int UserId, string PinCode, float Money);
+public record User(int UserId, string PinCode, double Money);
